requirements.txt
asgiref==3.3.1
block-disposable-email==2.0.0
boto3==1.16.25
botocore==1.19.25
braintree==4.5.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.44.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-disposable-email-checker==2.0.2
django-dotenv==1.4.2
django-hosts==4.0
django-phonenumber-field==5.0.0
django-private-storage==2.2.2
django-rest-auth==0.9.5
django-storages==1.10.1
django-tinymce==3.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
phonenumberslite==8.12.14
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
